Rework queue design.
All three queues are exposed to the application
and the application is free to use them in many
threads. Also only exposed the commands that
makes sense for the specific queue.

This paired with the timeline semaphores makes
it very easy to implement many different
architectures (Like a single queue like flow
like wgpu is operating or something with
async transfer and compute queues).